+++ packages/navigation/atlassian-switcher/examples/80-uc-jira-switcher-example.tsx
@@ -15,18 +15,7 @@ class JiraSwitcherExample extends React.Component {
   }
 
   openDrawer = () => {
-    mockEndpoints(
-      'jira',
-      originalMockData => {
-        return {
-          ...originalMockData,
-          LICENSE_INFORMATION_DATA: {
-            notUsedInMode: 'user-centric',
-          },
-        };
-      },
-      REQUEST_FAST,
-    );
+    mockEndpoints('jira', originalMockData => originalMockData, REQUEST_FAST);
     this.setState({
       isDrawerOpen: true,
     });
@@ -51,7 +40,6 @@ class JiraSwitcherExample extends React.Component {
           <AtlassianSwitcher
             product="jira"
             cloudId="some-cloud-id"
-            enableUserCentricProducts
             triggerXFlow={this.onTriggerXFlow}
           />
         </Drawer>